Bugfix for multi-byte characters.
write_object now uses the bytesize method to get the true size of the
string in number of bytes that will be written to the file. The decoding
in read_object now forces the encoding to UTF-8 (File.read returns
strings as ASCII-8BIT) so that it's decoded properly. Then, the unpack
method opens the output text file as UTF-8 so that we can write proper
characters to it.